Move `pkg/data/db` to `pkg/db`, make Open universal
